add new checkout_hash on payment requests
Flutterwave ensures secure payment on the client side by using a hashed value in a field called checkout_hash. This hash is created by encrypting some immutable values in your request. We compute the hash at runtime and compare it to the value that has been passed in your request to ensure that the payment is secure.